pyproject.toml
[project]
name = "talent-city-fa"
version = "0.1.0"
description = ""
authors = [
    {name = "Naisa",email = "r-r-y@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "black (>=25.1.0,<26.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (==4.0.1)",
    "redis (>=6.0.0,<7.0.0)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)",
    "celery (>=5.5.2,<6.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pillow (>=11.2.1,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
